feat: Implement authentication context and navigation structure
- Added AuthContext for managing user authentication state.
- Created AppStack for main application navigation including item screens.
- Developed AuthNavigator for handling authentication flows (login, account creation, password reset).
- Implemented RootNavigator to conditionally render AppStack or AuthNavigator based on authentication state.
- Added LoadingScreen component with animations and random phrases for user experience.
- Defined User type in types directory for future user-related functionalities.

+++ backend/src/main/java/com/example/schneider/trash_to_treasure/controller/UserController.java
@@ -37,19 +37,19 @@ public class UserController {
 
     //Request to find the user by id
     @GetMapping("/{id}")
-    public ResponseEntity<UserDTO.Response> findById(@PathVariable Integer id){
+    public ResponseEntity<UserDTO.Response> findById(@PathVariable Long id){
         return ResponseEntity.ok(userService.findById(id));
     }
     
     //Request to update the user by id
     @PutMapping("/{id}")
-    public ResponseEntity<UserDTO.Response> update(@PathVariable Integer id, @RequestBody @Valid UserDTO.Create userCreateDTO){
+    public ResponseEntity<UserDTO.Response> update(@PathVariable Long id, @RequestBody @Valid UserDTO.Create userCreateDTO){
         return ResponseEntity.ok(userService.update(id, userCreateDTO));
     }
 
     //Request to delete the user by id
     @DeleteMapping("/{id}")
-    public ResponseEntity<UserDTO.Response> delete(@PathVariable Integer id){
+    public ResponseEntity<UserDTO.Response> delete(@PathVariable Long id){
         userService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
